Limpieza del código de #Actividad.js
