+++ AppCalendarioMuroVertiz/components/ColorPicker.js
@@ -1,0 +1,40 @@
+import React from 'react';
+import { View, TouchableOpacity } from 'react-native';
+import styles from '../styles/CalendarioStyle';
+
+const COLORS = [
+  { name: 'Negro', value: '#000000' },
+  { name: 'Blanco', value: '#FFFFFF' },
+  { name: 'Gris', value: '#808080' },
+  { name: 'Rojo', value: '#FF0000' },
+  { name: 'Naranja', value: '#FFA500' },
+  { name: 'Amarillo', value: '#FFFF00' },
+  { name: 'Verde claro', value: '#90EE90' },
+  { name: 'Verde oscuro', value: '#006400' },
+  { name: 'Cyan', value: '#00FFFF' },
+  { name: 'Azul', value: '#0000FF' },
+  { name: 'Morado', value: '#800080' },
+  { name: 'Fuxia', value: '#FF00FF' },
+];
+
+const ColorPicker = ({ selectedColor, onSelect }) => (
+  <View style={styles.colorPickerContainer}>
+    {COLORS.map((color) => (
+      <TouchableOpacity
+        key={color.value}
+        style={[
+          styles.colorCircle,
+          {
+            backgroundColor: color.value,
+            borderColor: selectedColor === color.value ? '#333' : '#ccc',
+          },
+        ]}
+        onPress={() => onSelect(color.value)}
+      >
+        {selectedColor === color.value && <View style={styles.selectedCircle} />}
+      </TouchableOpacity>
+    ))}
+  </View>
+);
+
+export default ColorPicker;
